chores(x2a): split x2a-node services.ts
Refactoring only for better maintenance, no functional change.
