+++ workspaces/x2a/.changeset/crisp-hotels-stay.md
@@ -1,0 +1,5 @@
+---
+'@red-hat-developer-hub/backstage-plugin-x2a-node': patch
+---
+
+Internal change only - split x2a-node services.ts for better maintenance in the future.
