+++ workspaces/x2a/plugins/x2a-node/report.api.md
@@ -5,12 +5,12 @@
 ```ts
 
 import type { Artifact } from '@red-hat-developer-hub/backstage-plugin-x2a-common';
-import { BackstageCredentials } from '@backstage/backend-plugin-api';
-import { BackstageUserPrincipal } from '@backstage/backend-plugin-api';
+import type { BackstageCredentials } from '@backstage/backend-plugin-api';
+import type { BackstageUserPrincipal } from '@backstage/backend-plugin-api';
 import type { CatalogService } from '@backstage/plugin-catalog-node';
 import type { Job } from '@red-hat-developer-hub/backstage-plugin-x2a-common';
 import type { JobStatusEnum } from '@red-hat-developer-hub/backstage-plugin-x2a-common';
-import { LoggerService } from '@backstage/backend-plugin-api';
+import type { LoggerService } from '@backstage/backend-plugin-api';
 import type { MigrationPhase } from '@red-hat-developer-hub/backstage-plugin-x2a-common';
 import type { Module } from '@red-hat-developer-hub/backstage-plugin-x2a-common';
 import type { ModuleStatus } from '@red-hat-developer-hub/backstage-plugin-x2a-common';
