+++ workspaces/x2a/plugins/x2a-node/src/services/kubeServiceApi.ts
@@ -1,0 +1,36 @@
+/*
+ * Copyright Red Hat, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+import type { JobCreateParams, JobStatusInfo } from './types';
+
+/**
+ * Public API surface of the X2A Kubernetes service.
+ *
+ * The implementation lives in `x2a-backend` (`KubeService` class).
+ * Only methods whose signatures do not require `@kubernetes/client-node` types
+ * are included; this keeps the node-library free of heavy dependencies.
+ *
+ * @public
+ */
+export interface KubeServiceApi {
+  createJob(params: JobCreateParams): Promise<{ k8sJobName: string }>;
+  getJobStatus(k8sJobName: string): Promise<JobStatusInfo>;
+  getJobLogs(
+    k8sJobName: string,
+    streaming?: boolean,
+  ): Promise<string | NodeJS.ReadableStream>;
+  deleteJob(k8sJobName: string): Promise<void>;
+}
