+++ workspaces/x2a/plugins/x2a-node/src/services/reconcileJobDeps.ts
@@ -1,0 +1,30 @@
+/*
+ * Copyright Red Hat, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+import type { LoggerService } from '@backstage/backend-plugin-api';
+
+import type { X2ADatabaseServiceApi } from './X2ADatabaseService';
+import type { KubeServiceApi } from './kubeServiceApi';
+
+/**
+ * Dependencies for {@link reconcileJobStatus}.
+ * @public
+ */
+export interface ReconcileJobDeps {
+  kubeService: KubeServiceApi;
+  x2aDatabase: X2ADatabaseServiceApi;
+  logger: LoggerService;
+}
